+++ src/otx/algorithms/visual_prompting/adapters/openvino/model_wrappers.py
@@ -18,8 +18,8 @@ from typing import Any, Dict, Tuple
 
 import cv2
 import numpy as np
-from openvino.model_zoo.model_api.models import ImageModel
-from openvino.model_zoo.model_api.models.types import NumericalValue
+from openvino.model_api.models import ImageModel
+from openvino.model_api.models.types import NumericalValue
 
 from otx.algorithms.segmentation.adapters.openvino.model_wrappers.blur import (
     BlurSegmentation,