+++ tests/unit/algorithms/visual_prompting/adapters/openvino/test_model_wrappers.py
@@ -5,11 +5,11 @@
 #
 
 import numpy as np
-from openvino.model_zoo.model_api.models import SegmentationModel
+from openvino.model_api.models import SegmentationModel
 from otx.api.entities.label import LabelEntity
 import pytest
-from openvino.model_zoo.model_api.adapters.openvino_adapter import OpenvinoAdapter
-from openvino.model_zoo.model_api.models.types import NumericalValue
+from openvino.model_api.adapters.openvino_adapter import OpenvinoAdapter
+from openvino.model_api.models.types import NumericalValue
 from otx.algorithms.segmentation.adapters.openvino.model_wrappers.blur import (
     BlurSegmentation,
 )